Fix error pulling images
When attempting to pull images, there was an error about missing the
`.env.ngrok` file. This file is unnecessary for consuming the exercises
and is only needed when serving the repository.

This file however does contain a sensitive secret and must be Git
ignored.

By setting the environment file to `required: false`, this file is no
longer required.

Other references to environment files have been updated to this new
notation for consistency.